+++ crates/module-system/sov-test-utils/src/sequencer.rs
@@ -11,8 +11,8 @@ use sov_rollup_interface::node::batch_builder::BatchBuilder;
 use sov_rollup_interface::stf::StateTransitionFunction;
 use sov_rollup_interface::storage::HierarchicalStorageManager;
 use sov_sequencer::{
-    sequencer_rest_api_server, FairBatchBuilder, FairBatchBuilderConfig, GenericSequencerSpec,
-    Sequencer, SequencerDb, TxStatusManager,
+    FairBatchBuilder, FairBatchBuilderConfig, GenericSequencerSpec, Sequencer, SequencerDb,
+    TxStatusManager,
 };
 use sov_sequencer_json_client::Client;
 use sov_state::{DefaultStorageSpec, ProverStorage};
@@ -137,8 +137,7 @@ impl<B: BatchBuilder> TestSequencerSetup<B> {
 
         let (axum_addr, sequencer_axum_server) = {
             let addr = SocketAddr::from(([127, 0, 0, 1], 0));
-            let router =
-                sequencer_rest_api_server("/sequencer").with_state::<()>(sequencer.clone());
+            let router = sequencer.rest_api_server("/sequencer");
 
             let handle = axum_server::Handle::new();
             let handle1 = handle.clone();
@@ -252,8 +251,7 @@ impl TestSequencerSetup<TestFairBatchBuilder> {
 
         let (axum_addr, sequencer_axum_server) = {
             let addr = SocketAddr::from(([127, 0, 0, 1], 0));
-            let router =
-                sequencer_rest_api_server("/sequencer").with_state::<()>(sequencer.clone());
+            let router = sequencer.rest_api_server("/sequencer");
 
             let handle = axum_server::Handle::new();
             let handle1 = handle.clone();