+++ examples/demo-rollup/src/celestia_rollup.rs
@@ -3,7 +3,7 @@ use backon::ExponentialBuilder;
 use demo_stf::authentication::ModAuth;
 use demo_stf::runtime::Runtime;
 use sov_celestia_adapter::verifier::{CelestiaSpec, CelestiaVerifier, RollupParams};
-use sov_celestia_adapter::{CelestiaConfig, CelestiaService};
+use sov_celestia_adapter::CelestiaService;
 use sov_db::ledger_db::LedgerDb;
 use sov_db::storage_manager::NativeStorageManager;
 use sov_kernels::basic::BasicKernel;
@@ -16,7 +16,7 @@ use sov_modules_rollup_blueprint::{FullNodeBlueprint, RollupBlueprint, WalletBlu
 use sov_modules_stf_blueprint::{RuntimeEndpoints, StfBlueprint};
 use sov_risc0_adapter::host::Risc0Host;
 use sov_risc0_adapter::Risc0Verifier;
-use sov_rollup_interface::node::da::DaServiceWithRetries;
+use sov_rollup_interface::node::da::{DaService, DaServiceWithRetries};
 use sov_rollup_interface::zk::aggregated_proof::CodeCommitment;
 use sov_rollup_interface::zk::Zkvm;
 use sov_sequencer::{FairBatchBuilderConfig, SequencerDb};
@@ -45,7 +45,6 @@ where
 #[async_trait]
 impl FullNodeBlueprint<Native> for CelestiaDemoRollup<Native> {
     type DaService = DaServiceWithRetries<CelestiaService>;
-    type DaConfig = CelestiaConfig;
 
     type InnerZkvmHost = Risc0Host<'static>;
     type OuterZkvmHost = MockZkvm;
@@ -86,7 +85,7 @@ impl FullNodeBlueprint<Native> for CelestiaDemoRollup<Native> {
         da_service: &Self::DaService,
         rollup_config: &RollupConfig<
             <Self::Spec as Spec>::Address,
-            Self::DaConfig,
+            <Self::DaService as DaService>::Config,
             FairBatchBuilderConfig<Self::DaSpec>,
         >,
     ) -> anyhow::Result<RuntimeEndpoints> {
@@ -117,7 +116,7 @@ impl FullNodeBlueprint<Native> for CelestiaDemoRollup<Native> {
         &self,
         rollup_config: &RollupConfig<
             <Self::Spec as Spec>::Address,
-            Self::DaConfig,
+            <Self::DaService as DaService>::Config,
             FairBatchBuilderConfig<Self::DaSpec>,
         >,
     ) -> Self::DaService {
@@ -141,7 +140,7 @@ impl FullNodeBlueprint<Native> for CelestiaDemoRollup<Native> {
         prover_config: RollupProverConfig,
         rollup_config: &RollupConfig<
             <Self::Spec as Spec>::Address,
-            Self::DaConfig,
+            <Self::DaService as DaService>::Config,
             FairBatchBuilderConfig<Self::DaSpec>,
         >,
         _da_service: &Self::DaService,
@@ -173,7 +172,7 @@ impl FullNodeBlueprint<Native> for CelestiaDemoRollup<Native> {
         &self,
         rollup_config: &RollupConfig<
             <Self::Spec as Spec>::Address,
-            Self::DaConfig,
+            <Self::DaService as DaService>::Config,
             FairBatchBuilderConfig<Self::DaSpec>,
         >,
     ) -> anyhow::Result<Self::StorageManager> {